Add TLSA record (not very tested)
As with CAA, I only checked that dig appeared to like it.
dig did split up the value, and in a different spot than examples would suggest, but I don't know if bind would do it differently or if it is a problem. I suspect it is simply display formatting