Add support for sub /24 delegations.
Some support for sub /124 is present as well, but incomplete.

Although it is normally better to delegate each IP individually, some
DNS servers make it a pain as you need a separate file and config for
each zone. This eases interop with such servers and services that use
such servers.